major overhaul refactor RecSys module
