+++ orange_cb_recsys/recsys/__init__.py
@@ -1,3 +1,1 @@
-from .config import RecSysConfig, ContentBasedConfig, GraphBasedConfig
-from .recsys import RecSys
-from .algorithm import Algorithm
+
