+++ orange_cb_recsys/recsys/content_based_algorithm/__init__.py
@@ -1,7 +1,4 @@
 from .content_based_algorithm import ContentBasedAlgorithm
-from .centroid_vector import CentroidVector
-from .classifier import ClassifierRecommender
-from .index_query import IndexQuery
-from .similarities import CosineSimilarity
-import pandas as pd
-from orange_cb_recsys.utils.const import logger
+from orange_cb_recsys.recsys.content_based_algorithm.classifier.classifier_recommender import ClassifierRecommender
+from .index_query.index_query import IndexQuery
+from orange_cb_recsys.recsys.content_based_algorithm.centroid_vector.similarities import CosineSimilarity
