+++ orange_cb_recsys/recsys/content_based_algorithm/content_based_algorithm.py
@@ -10,7 +10,6 @@ import numpy as np
 from orange_cb_recsys.recsys.algorithm import Algorithm
 
 from orange_cb_recsys.content_analyzer.content_representation.content import Content
-from orange_cb_recsys.recsys.content_based_algorithm.exceptions import Handler_EmptyFrame
 
 
 class ContentBasedAlgorithm(Algorithm):
@@ -36,25 +35,6 @@ class ContentBasedAlgorithm(Algorithm):
         self.item_field: dict = self._bracket_representation(item_field)
         self.threshold: float = threshold
         self.__transformer = DictVectorizer(sparse=True, sort=False)
-        self.users_directory: str = None
-        self.items_directory: str = None
-        self.ratings: pd.DataFrame = None
-
-    def initialize(self, ratings: pd.DataFrame, items_directory: str, users_directory: str = None):
-        """
-        Method to call right after the instantiation of the algorithm.
-
-        Its task is to pass important parameters to the algorithm, needed to calculate the rank
-        or the predictions
-
-        Args:
-            ratings (pd.DataFrame): DataFrame containing ratings that users assigned to items
-            items_directory (str): path of the directory where the items are stored
-            users_directory (str): path of the directory where the users are stored
-        """
-        self.ratings = ratings
-        self.items_directory = items_directory
-        self.users_directory = users_directory
 
     @staticmethod
     def _bracket_representation(item_field: dict):
@@ -79,6 +59,10 @@ class ContentBasedAlgorithm(Algorithm):
                 item_field[field] = [item_field[field]]
 
         return item_field
+
+    @staticmethod
+    def _calc_mean_user_threshold(user_ratings: pd.DataFrame):
+        return user_ratings['score'].mean()
 
     def extract_features_item(self, item: Content):
         """
@@ -109,19 +93,6 @@ class ContentBasedAlgorithm(Algorithm):
                     )
 
         return item_bag_list
-
-    def assign_class(self, score: float):
-        """
-        Function that assigns a class for the item
-
-        Args:
-            score (float): the rating of the user
-
-        Returns:
-            Returns 1 if the score passed is >= than the threshold, 0 otherwise
-            The threshold is the parameter specified in the constructor
-        """
-        return 1 if score >= self.threshold else 0
 
     def fuse_representations(self, X: list):
         """
@@ -276,63 +247,63 @@ class ContentBasedAlgorithm(Algorithm):
         """
         raise NotImplementedError
 
-    @Handler_EmptyFrame
-    def fit_predict(self, user_id: str, filter_list: List[str] = None):
-        """
-        Method used to predict the rating of the user passed for all unrated items or for the items passed
-        in the filter_list parameter.
-
-        The method fits the algorithm and then calculates the prediction
-
-        Args:
-            user_id (str): user_id of the user
-            filter_list (list): list of the items to rank, if None all unrated items will be used to
-                calculate the rank
-        Returns:
-            pd.DataFrame: DataFrame containing one column with the items name,
-                one column with the rating predicted
-        """
-        # Extracts ratings of the user
-        user_ratings = self.ratings[self.ratings['from_id'] == user_id]
-
-        # Process rated items
-        self.process_rated(user_ratings, self.items_directory)
-
-        # Fit
-        self.fit()
-
-        # Predict
-        prediction = self.predict(user_ratings, self.items_directory, filter_list)
-        return prediction
-
-    @Handler_EmptyFrame
-    def fit_rank(self, user_id: str, recs_number: int = None, filter_list: List[str] = None):
-        """
-        Method used to rank for a particular user all unrated items or the items specified in
-        the filter_list parameter.
-
-        The method fits the algorithm and then calculates the rank.
-
-        If the recs_number is specified, then the rank will contain the top-n items for the user.
-
-        Args:
-            user_id (str): user_id of the user
-            recs_number (int): number of the top items that will be present in the ranking
-            filter_list (list): list of the items to rank, if None all unrated items will be used to
-                calculate the rank
-        Returns:
-            pd.DataFrame: DataFrame containing one column with the items name,
-                one column with the rating predicted, sorted in descending order by the 'rating' column
-        """
-        # Extracts ratings of the user
-        user_ratings = self.ratings[self.ratings['from_id'] == user_id]
-
-        # Process rated items
-        self.process_rated(user_ratings, self.items_directory)
-
-        # Fit
-        self.fit()
-
-        # Rank
-        rank = self.rank(user_ratings, self.items_directory, recs_number, filter_list)
-        return rank
+    # @Handler_EmptyFrame
+    # def fit_predict(self, user_id: str, filter_list: List[str] = None):
+    #     """
+    #     Method used to predict the rating of the user passed for all unrated items or for the items passed
+    #     in the filter_list parameter.
+    #
+    #     The method fits the algorithm and then calculates the prediction
+    #
+    #     Args:
+    #         user_id (str): user_id of the user
+    #         filter_list (list): list of the items to rank, if None all unrated items will be used to
+    #             calculate the rank
+    #     Returns:
+    #         pd.DataFrame: DataFrame containing one column with the items name,
+    #             one column with the rating predicted
+    #     """
+    #     # Extracts ratings of the user
+    #     user_ratings = self.ratings[self.ratings['from_id'] == user_id]
+    #
+    #     # Process rated items
+    #     self.process_rated(user_ratings, self.items_directory)
+    #
+    #     # Fit
+    #     self.fit()
+    #
+    #     # Predict
+    #     prediction = self.predict(user_ratings, self.items_directory, filter_list)
+    #     return prediction
+    #
+    # @Handler_EmptyFrame
+    # def fit_rank(self, user_id: str, recs_number: int = None, filter_list: List[str] = None):
+    #     """
+    #     Method used to rank for a particular user all unrated items or the items specified in
+    #     the filter_list parameter.
+    #
+    #     The method fits the algorithm and then calculates the rank.
+    #
+    #     If the recs_number is specified, then the rank will contain the top-n items for the user.
+    #
+    #     Args:
+    #         user_id (str): user_id of the user
+    #         recs_number (int): number of the top items that will be present in the ranking
+    #         filter_list (list): list of the items to rank, if None all unrated items will be used to
+    #             calculate the rank
+    #     Returns:
+    #         pd.DataFrame: DataFrame containing one column with the items name,
+    #             one column with the rating predicted, sorted in descending order by the 'rating' column
+    #     """
+    #     # Extracts ratings of the user
+    #     user_ratings = self.ratings[self.ratings['from_id'] == user_id]
+    #
+    #     # Process rated items
+    #     self.process_rated(user_ratings, self.items_directory)
+    #
+    #     # Fit
+    #     self.fit()
+    #
+    #     # Rank
+    #     rank = self.rank(user_ratings, self.items_directory, recs_number, filter_list)
+    #     return rank
