+++ orange_cb_recsys/recsys/graphs/__init__.py
@@ -1,4 +1,4 @@
 from .graph import Graph, BipartiteGraph, TripartiteGraph, FullGraph
-from .bipartite_graphs import NXBipartiteGraph
-from .tripartite_graphs import NXTripartiteGraph
-from .full_graphs import NXFullGraph
+from .nx_bipartite_graphs import NXBipartiteGraph
+from .nx_tripartite_graphs import NXTripartiteGraph
+from .nx_full_graphs import NXFullGraph
