+++ orange_cb_recsys/recsys/graphs/graph_metrics.py
@@ -1,0 +1,16 @@
+from abc import ABC, abstractmethod
+
+
+class GraphMetrics(ABC):
+
+    @abstractmethod
+    def degree_centrality(self):
+        raise NotImplementedError
+
+    @abstractmethod
+    def closeness_centrality(self):
+        raise NotImplementedError
+
+    @abstractmethod
+    def dispersion(self):
+        raise NotImplementedError
